sprint-4/step-1 - ApiService extracted
